requirements.txt
gradio==4.44.0
torch
transformers
accelerate
sentencepiece
Pillow

